backend/requirements.txt
# Core Framework (FastAPI + Uvicorn)
fastapi==0.128.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20  # Critical for video file uploads (UploadFile)

# AI & Intelligence (Gemini)
google-genai==1.0.0

# Browser Automation (The Hands)
playwright==1.57.0

# Data Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utilities
aiofiles==24.1.0          # Async file operations (optional but recommended for video handling)
httpx==0.28.1             # For any internal async HTTP requests
yt-dlp==2024.12.23        # For downloading YouTube videos
opencv-python==4.10.0.84  # For video frame extraction